Combine Parquet testing headers
