+++ cpp/src/parquet/column_writer-test.cc
@@ -26,7 +26,6 @@
 #include "parquet/metadata.h"
 #include "parquet/properties.h"
 #include "parquet/statistics.h"
-#include "parquet/test-specialization.h"
 #include "parquet/test-util.h"
 #include "parquet/thrift.h"
 #include "parquet/types.h"
