+++ cpp/src/parquet/file-deserialize-test.cc
@@ -25,10 +25,10 @@
 #include "parquet/column_reader.h"
 #include "parquet/exception.h"
 #include "parquet/file_reader.h"
+#include "parquet/test-util.h"
 #include "parquet/thrift.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/test-common.h"
 
 #include "arrow/io/memory.h"
 #include "arrow/status.h"
