+++ cpp/src/parquet/statistics-test.cc
@@ -30,7 +30,6 @@
 #include "parquet/file_writer.h"
 #include "parquet/schema.h"
 #include "parquet/statistics.h"
-#include "parquet/test-specialization.h"
 #include "parquet/test-util.h"
 #include "parquet/thrift.h"
 #include "parquet/types.h"
