+++ cpp/src/parquet/util/memory-test.cc
@@ -24,8 +24,8 @@
 #include <gtest/gtest.h>
 
 #include "parquet/exception.h"
+#include "parquet/test-util.h"
 #include "parquet/util/memory.h"
-#include "parquet/util/test-common.h"
 
 using arrow::default_memory_pool;
 using arrow::MemoryPool;
